AP-40: Change time accumulator to 32 bit rather than 64 bit

+++ libraries/AP_OpticalFlow/AP_OpticalFlow_MotorPod.h
@@ -28,7 +28,7 @@ private:
     struct GyroAccum {
         float x;
         float y;
-        uint64_t t;
+        uint32_t t;
     } gyro_accum;
 
     uint32_t lastUpdate_us;
